pyproject.toml
[tool.poetry]
name = "homework4"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
python-multipart = "^0.0.5"
rq = "^1.7.0"
requests = "^2.25.1"
Pillow = "^8.1.2"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
autoflake = "^1.4"
isort = "^5.7.0"
black = "^20.8b1"
unify = "^0.5"
flake8 = "^3.8.4"
pylint = "^2.7.1"
mypy = "^0.812"
pytest-deadfixtures = "^2.2.1"
fakeredis = "^1.5.0"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
